Add functions and dialog form
I used the dialog form with action=dialog, it was easy to get all the
user input information using the elements method to organize all of them.

The function addBookToLibrary adds a book and writes to the screen the
book adding the necessary html elements. The button is identified with
an custom atribute to identify the card when it is called. the for loop
that iterates the book object and adds the content to the html tags is
overly complicated this exemple, but could be useful for large projects.